Separate AbortHandler
Move the handler into a separate file and create a simple unit test to
verify it.
