+++ collector/collector.cpp
@@ -37,8 +37,6 @@ You should have received a copy of the GNU General Public License along with thi
 extern "C" {
 #include <assert.h>
 #include <cap-ng.h>
-#include <cxxabi.h>
-#include <dlfcn.h>
 #include <execinfo.h>
 #include <fcntl.h>
 #include <stdio.h>
@@ -50,6 +48,7 @@ extern "C" {
 #include <sys/syscall.h>
 #include <sys/types.h>
 }
+#include "AbortHandler.h"
 #include "CollectorArgs.h"
 #include "CollectorService.h"
 #include "CollectorStatsExporter.h"
@@ -81,89 +80,6 @@ ShutdownHandler(int signum) {
   // Only set the control variable; the collector service will take care of the rest.
   g_signum.store(signum);
   g_control.store(ControlValue::STOP_COLLECTOR);
-}
-
-static void WriteBuffer(char* buffer, int num_bytes, int buffer_size) {
-  if (num_bytes > buffer_size) {
-    // the output was truncated, to keep it readable add a new line
-    buffer[buffer_size - 1] = '\n';
-    write(STDERR_FILENO, buffer, buffer_size);
-  } else {
-    write(STDERR_FILENO, buffer, num_bytes);
-  }
-}
-
-static void AbortHandler(int signum) {
-  // Write a stacktrace to stderr. Since the abort handler could be called
-  // after SIGSEGV or some other emergency situations:
-  // * we have to be modest and use little memory (in case if the original
-  //   reason was OOM)
-  // * as the original implementation noted, use reentrant functions
-  //
-  // The latter one was refering to snprintf, although it's not clear if it is
-  // reentrant. Libc docs says functions using I/O streams are potentially
-  // non-reentrant [1], and it seems under the hood snprintf uses streams.
-  //
-  // [1]: https://www.gnu.org/savannah-checkouts/gnu/libc/manual/html_node/Nonreentrancy.html
-
-  void* buffer[32];  // Addresses buffer
-
-  char message_buffer[256];  // Actual text buffer for a single stack trace
-                             // address. Some templated signatures could be
-                             // particularly long.
-  int num_bytes;
-
-  size_t n_frames = backtrace(buffer, 32);
-  size_t max_frames = sizeof(buffer) / sizeof(buffer[0]);
-  int message_buffer_size = sizeof(message_buffer);
-
-  for (size_t i = 0; i < n_frames; i++) {
-    Dl_info info;
-
-    if (dladdr(buffer[i], &info)) {
-      // Try to demangle the stacktrace
-      char* demangled = NULL;
-      int status;
-      size_t shift_from_symbol;
-
-      // NULL as a second argument for the output buffer will make it allocate
-      // new region of memory
-      demangled = abi::__cxa_demangle(info.dli_sname, NULL, NULL, &status);
-      shift_from_symbol = (char*)buffer[i] - (char*)info.dli_saddr;
-
-      // The pattern is:
-      // - file name of the object
-      // - demangled or original name of the obect (or "(null)" if NULL)
-      // - address of the object
-      // - distance to the closest symbol
-      num_bytes = snprintf(message_buffer, message_buffer_size,
-                           "%s %s %p + %zd\n", info.dli_fname,
-                           status == 0 ? demangled : info.dli_sname, buffer[i],
-                           info.dli_saddr != NULL ? shift_from_symbol : 0);
-      free(demangled);
-      WriteBuffer(message_buffer, num_bytes, message_buffer_size);
-
-    } else {
-      // Failed to translate the address into symbolic information,
-      // at least print the raw address value.
-      num_bytes = snprintf(message_buffer, message_buffer_size, "%p\n", buffer[i]);
-      WriteBuffer(message_buffer, num_bytes, message_buffer_size);
-    }
-  }
-
-  if (n_frames == max_frames)
-    write(STDERR_FILENO, "[truncated]\n", 14);
-
-  // Write a message to stderr using only reentrant functions.
-  num_bytes = snprintf(message_buffer, message_buffer_size,
-                       "Caught signal %d (%s): %s\n",
-                       signum, SignalName(signum), strsignal(signum));
-  WriteBuffer(message_buffer, num_bytes, message_buffer_size);
-
-  // Re-raise the signal (this time routing it to the default handler) to make
-  // sure we get the correct exit code.
-  signal(signum, SIG_DFL);
-  raise(signum);
 }
 
 // creates a GRPC channel, using the tls configuration provided from the args.
